+++ modules/refactorkit-java/src/main/kotlin/org/refactorkit/java/JavaProjectScanner.kt
@@ -47,6 +47,8 @@ class JavaProjectScanner {
             workspace = Workspace(normalizedRoot),
             modules = modules,
             files = files,
+            sourceExtensions = setOf("java"),
+            ignoredDirectories = emptySet(),
         )
     }